Use Field class [part 3]
