+++ src/form/index.ts
@@ -1,2 +1,3 @@
-export * from './form';
-export * from './form-store';
+export * from './Form';
+export * from './Field';
+export * from './store';
